show latitude and long titude as external state
