+++ README.md
@@ -3,5 +3,15 @@
 This project was bootstrapped with [Create React App](https://dev.to/maj07/tutorial-react-leaflet-d65).
 
 ## Link to Tutorial
+[React Leaflet](https://react-leaflet.js.org/docs/example-external-state/)
 [DEV React Leaflet Marker Tutorial](https://dev.to/maj07/tutorial-react-leaflet-d65).
 [Log Rocket](https://blog.logrocket.com/react-leaflet-tutorial/)
+
+### App.css
+Has to set in `App.css` or Map will not render 
+```
+.leaflet-container {
+  width: 100%;
+  height: 100vh;
+}
+```
