--- src/TestApp.js
@@ -1,60 +1,0 @@
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { MapContainer, TileLayer } from 'react-leaflet';
-
-const latitude = 13.7554
-const longitude = 100.4932
-const center = [latitude, longitude]
-const zoom = 17
-
-function DisplayPosition({ map }) {
-  const [position, setPosition] = useState(() => map.getCenter())
-
-  const onClick = useCallback(() => {
-    map.setView(center, zoom)
-  }, [map])
-
-  const onMove = useCallback(() => {
-    setPosition(map.getCenter())
-  }, [map])
-
-  useEffect(() => {
-    map.on('move', onMove)
-    return () => {
-      map.off('move', onMove)
-    }
-  }, [map, onMove])
-
-  return (
-    <p>
-      latitude: {position.lat.toFixed(4)}, longitude: {position.lng.toFixed(4)}{' '}
-      <button onClick={onClick}>reset</button>
-    </p>
-  )
-}
-
-function TestApp() {
-  const [map, setMap] = useState()
-  const displayMap = useMemo(
-    () => (
-      <MapContainer
-        center={[latitude, longitude]}
-        zoom={zoom}
-        scrollWheelZoom={false}
-        ref={setMap}>
-        <TileLayer
-          attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
-          url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-        />
-      </MapContainer>
-    ),
-    [],
-  )
-  return (
-    <div>
-      {map ? <DisplayPosition map={map} /> : null}
-      {displayMap}
-    </div>
-  )
-}
-
-export default TestApp;
